refactor bookmark option scroll event
