+++ src/client/component/contentScript/BookmarkSidebar/BookmarkList/BookmarkList.module.css
@@ -2,3 +2,38 @@
     list-style: none;
     padding: 0;
 }
+
+.listScroll {
+    overflow-x: hidden;
+    overflow-y: scroll;
+    height: calc(100% - 40px);
+}
+
+.listScroll::-webkit-scrollbar {
+    width: 12px;
+}
+
+.listScroll::-webkit-scrollbar-thumb {
+    visibility: hidden;
+    border-radius: 10px;
+    background-color: #bbbbbb;
+    background-clip: padding-box;
+    border: 3px solid transparent;
+}
+
+.listScroll::-webkit-scrollbar-thumb:hover {
+    background-color: #999999;
+}
+
+.listScroll::-webkit-scrollbar-thumb:active {
+    background-color: #777777;
+}
+
+.listScroll:hover::-webkit-scrollbar-thumb {
+    visibility: visible;
+}
+
+.listScroll::-webkit-scrollbar-track {
+    background-color: transparent;
+    border-radius: 10px;
+}
