+++ src/client/component/contentScript/BookmarkSidebar/BookmarkSidebar.module.css
@@ -15,41 +15,6 @@
     color: rgba(0, 0, 0, 0);
 }
 
-.listScroll {
-    overflow-x: hidden;
-    overflow-y: scroll;
-    height: calc(100% - 40px);
-}
-
-.listScroll::-webkit-scrollbar {
-    width: 12px;
-}
-
-.listScroll::-webkit-scrollbar-thumb {
-    visibility: hidden;
-    border-radius: 10px;
-    background-color: #bbbbbb;
-    background-clip: padding-box;
-    border: 3px solid transparent;
-}
-
-.listScroll::-webkit-scrollbar-thumb:hover {
-    background-color: #999999;
-}
-
-.listScroll::-webkit-scrollbar-thumb:active {
-    background-color: #777777;
-}
-
-.listScroll:hover::-webkit-scrollbar-thumb {
-    visibility: visible;
-}
-
-.listScroll::-webkit-scrollbar-track {
-    background-color: transparent;
-    border-radius: 10px;
-}
-
 .background {
     display: block;
     opacity: 0;
